mqtt/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
click==8.1.8
fastapi==0.115.6
h11==0.14.0
idna==3.10
influxdb-client==1.48.0
paho-mqtt==2.1.0
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
reactivex==4.0.4
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
